requirements.txt
BeautifulSoup==3.2.1
docopt==0.6.2
lxml==3.7.3
python-docx==0.8.6
requests==2.13.0
